main/backend/requirements.txt
flask
flask-cors
opencv-python
onnxruntime
numpy
python-dotenv
ultralytics
Pillow
openai